+++ src/typey-boi/containers/TypeyBoi.tsx
@@ -4,6 +4,7 @@ import { LoadText } from 'components/LoadText'
 import { CompletedParagraphs } from 'components/paragraphs/CompletedParagraphs'
 import { CurrentParagraph } from 'components/paragraphs/CurrentParagraph'
 import { FutureParagraphs } from 'components/paragraphs/FutureParagraphs'
+import { Statistics } from 'components/Statistics'
 
 export function TypeyBoi(): React.ReactElement {
   return (
@@ -14,6 +15,7 @@ export function TypeyBoi(): React.ReactElement {
         <CurrentParagraph />
         <FutureParagraphs />
       </div>
+      <Statistics />
       <LoadText />
     </>
   )